Refactor Started and Credential Error exception added
